requirements.txt
aniso8601==1.1.0
backports-abc==0.4
boto3==1.3.1
botocore==1.4.58
cachetools==2.0.1
cassandra-driver==3.4.1
click==6.7
coverage==4.0.3
docutils==0.12
elasticsearch==2.3.0
facebook-sdk==2.0.0
fake-factory==0.2
Flask==0.12.2
Flask-RESTful==0.3.5
gevent==1.1.1
gitdb2==2.0.0
greenlet==0.4.10
hiredis==0.2.0
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.0
kruger==0.9.24
MarkupSafe==1.0
nose==1.3.7
oauth2==1.9.0.post1
passlib==1.6.5
pkginfo==1.4.1
PyJWT==1.4.0
python-dateutil==2.5.3
pytz==2016.4
PyYAML==3.11
redis==2.10.5
requests==2.9.1
requests-toolbelt==0.8.0
retrying==1.3.3
s3transfer==0.1.5
six==1.10.0
smmap2==2.0.1
tornado==4.3
tqdm==4.15.0
twine==1.9.1
ujson==1.33
urllib3==1.18
virtualenv==15.0.3
Werkzeug==0.12.2
